Callback for `from` (#46)
* Callback for `from`

Added a callback for `from`, which takes file as an argument. Allows the user to tailor the search string according to the filename.

* Minor Changes to Callback for `from`

Additional comments to explain the callback forms.
Reversed minor version bump.

* Changes to documentation for `to` callback

Reverted to original `to` callback example.
Added a separate example example for using the `file` argument.